Initial commit.
This includes a Packer definition to build a Debian Buster Vagrant box
suitable for the VirtualBox provider.

The image is quite minimal at this point. Additional variants can be
created later by providing additional scripts (in the scripts/
directory) and referencing them in the variables/ directory.

A sample Vagrantfile is also given to show how to specify the
debian/debian username and password used insted of the usual
vagrant/vagrant used in Vagrant boxes.